implement closures and understand iterators

+++ src/libs/factorial.rs
@@ -1,4 +1,4 @@
-pub fn factorial(number: u16) -> u64 {
+pub fn factorial(number: u32) -> u64 {
   let mut fact: u64 = 1;
   for n in 2..(number + 1) {
     fact = fact * (n as u64);
